requirements.txt
# Core Data Science Libraries
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3
tensorflow==2.15.0
keras==2.15.0

# Time Series
prophet==1.1.5
statsmodels==0.14.1

# Big Data (Spark)
pyspark==3.5.0

# Visualization
streamlit==1.29.0
streamlit-folium==0.15.1
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0
folium==0.15.1

# Geospatial & Network Analysis
networkx==3.2.1
geopandas==0.14.1
shapely==2.0.2

# Optimization
scipy>=1.11.4
pulp==2.7.0

# Utilities
pyyaml==6.0.1
loguru==0.7.2
python-dotenv==1.0.0
tqdm==4.66.1

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1
notebook==7.0.6

# Testing & Quality
pytest==7.4.3
black==23.12.1
flake8==6.1.0
